+++ auction-ui/src/routes/boards/Board-admin.component.jsx
@@ -1,8 +1,8 @@
 import React, { Component } from "react";
 
-import UserService from "../services/user.service";
+import UserService from "../../services/user.service";
 
-export default class BoardUser extends Component {
+export default class BoardAdmin extends Component {
     constructor(props) {
         super(props);
 
@@ -12,7 +12,7 @@ export default class BoardUser extends Component {
     }
 
     componentDidMount() {
-        UserService.getUserBoard().then(
+        UserService.getAdminBoard().then(
             response => {
                 this.setState({
                     content: response.data
