+++ auction-ui/src/routes/boards/Board-user.component.jsx
@@ -1,8 +1,8 @@
 import React, { Component } from "react";
 
-import UserService from "../services/user.service";
+import UserService from "../../services/user.service";
 
-export default class Home extends Component {
+export default class BoardUser extends Component {
     constructor(props) {
         super(props);
 
@@ -12,7 +12,7 @@ export default class Home extends Component {
     }
 
     componentDidMount() {
-        UserService.getPublicContent().then(
+        UserService.getUserBoard().then(
             response => {
                 this.setState({
                     content: response.data
@@ -21,7 +21,9 @@ export default class Home extends Component {
             error => {
                 this.setState({
                     content:
-                        (error.response && error.response.data) ||
+                        (error.response &&
+                            error.response.data &&
+                            error.response.data.message) ||
                         error.message ||
                         error.toString()
                 });
@@ -31,7 +33,7 @@ export default class Home extends Component {
 
     render() {
         return (
-            <div className="container" >
+            <div className="container">
                 <header className="jumbotron">
                     <h3>{this.state.content}</h3>
                 </header>
