+++ auction-ui/src/routes/home/Home.component.jsx
@@ -1,8 +1,8 @@
-import React, { Component } from "react";
+import { Component } from "react";
 
-import UserService from "../services/user.service";
+import UserService from "../../services/user.service";
 
-export default class BoardAdmin extends Component {
+export default class Home extends Component {
     constructor(props) {
         super(props);
 
@@ -12,7 +12,7 @@ export default class BoardAdmin extends Component {
     }
 
     componentDidMount() {
-        UserService.getAdminBoard().then(
+        UserService.getPublicContent().then(
             response => {
                 this.setState({
                     content: response.data
@@ -21,9 +21,7 @@ export default class BoardAdmin extends Component {
             error => {
                 this.setState({
                     content:
-                        (error.response &&
-                            error.response.data &&
-                            error.response.data.message) ||
+                        (error.response && error.response.data) ||
                         error.message ||
                         error.toString()
                 });
@@ -33,7 +31,7 @@ export default class BoardAdmin extends Component {
 
     render() {
         return (
-            <div className="container">
+            <div className="container" >
                 <header className="jumbotron">
                     <h3>{this.state.content}</h3>
                 </header>
