+++ auction-ui/src/routes/login-signup/Login.component.jsx
@@ -3,9 +3,9 @@ import Form from "react-validation/build/form";
 import Input from "react-validation/build/input";
 import CheckButton from "react-validation/build/button";
 
-import AuthService from "../services/auth.service";
+import AuthService from "../../services/auth.service";
 
-import {withRouter} from '../common/with-router';
+import {withRouter} from './components/with-router';
 
 const required = value => {
     if (!value) {
