+++ auction-ui/src/routes/login-signup/Signup.component.jsx
@@ -4,7 +4,7 @@ import Input from "react-validation/build/input";
 import CheckButton from "react-validation/build/button";
 import isEmail from "validator/es/lib/isEmail.js";
 
-import AuthService from "../services/auth.service";
+import AuthService from "../../services/auth.service";
 
 
 const required = value => {
