+++ auction-ui/src/services/auth.service.jsx
@@ -4,7 +4,7 @@ const API_URL = "http://localhost:8080/api/auth/";
 
 class AuthService {
     login(username, password) {
-        debugger
+        
         return axios
             .post(API_URL + "signin", {
                 username,
